Update ufo-architecture.md
- If running the current .user-flowrc.json config it does not run and asks to set up the URL.

- user-flow directory was changed to user-flows as it may and probably will contain multiple. 

- Comment/Question.
There is a typescript error when passing the page to a Ufo, I have been passing the ctx but I won't work in a Ufo that uses another Ufo. 
The error in question is:
```ts
Argument of type '{ page: Page; }' is not assignable to parameter of type 'UserFlowContext'.
  Type '{ page: Page; }' is missing the following properties from type 'UserFlowContext': browser, flow, collectOptions
```
And the code in question is this:
```
export declare class Ufo {
    protected page: Page;
    constructor({ page }: UserFlowContext);
}
```
We can either grow the class to expose ctx `UserFlowContext` or reduce it by only passing the page. 
Or maybe there is a smarter way of doing this, some typescript magic.